Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I think it's ok to back-port this to Gingerbread. The changes do not seem to affect the ARM VM, but I'd like bill and fadden's confirmation on this.

Note that there is another depending patch to apply on top of this.

Patch-set: 1
Reviewer: David Turner <1000411@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1
Label: Verified=0