Fixing a deadlock in the HeapWorker thread
When the HeapWorker tries to lock the heapWorkerLock while
a GC is ongoing, there was a small risk that we got a
deadlock.

The HeapWorker thread would take the heapWorkerLock and then
become suspended, waiting for the GC thread to wake it up. But
the GC thread would get stuck, trying to get the heapWorkerLock
(held by the HeapWorker thread).

The fix is that the HeapWorker releases the heapWorkerLock
before it is suspended.

Change-Id: I736ee1e281ae993066a1d31d16a27f1c7ad370d6