Update patch set 1
Patch Set 1:

Here are two things which should be discussed:

1. I've added a commented block of code which tries to free
memory which was over-allocated before. We can remove the block if the consensus is that not enough memory is wasted here to warrant the additional complexity of reallocation.

2. After the change there's a rather big commented block of code which in the past tried to do the same as this patch. I'd be in favor of removing this block altogether, because a) it is superseded by this patch and b) the comment 

         * UPDATE: this makes ReferenceType.Interfaces difficult to implement,
         * because it wants to return just the interfaces declared to be
         * implemented directly by the class.  I'm excluding this code for now.

is not appropriate anyways: The implementation of ReferenceType.Interfaces didn't work as intended as witness in [1].

I've uploaded another patch [2] which fixes this other issue so that IMO it is safe to remove the whole excluded code block now. If you agree I'll update the patch accordingly.

[1] http://code.google.com/p/android/issues/detail?id=21422
[2] https://android-review.googlesource.com/#/c/30910/

Patch-set: 1
Label: Code-Review=0