Update patch set 4
Patch Set 4:

The missing lines that JBQ mentioned seem to have been moved to the top of the file.  Whoever submits this will have to be prepared to handle the merge conflict appropriately, moving the changes to the new location of the code in question.

Other than that, everything merges and builds fine in both AOSP and Google's internal master.

Patch-set: 4