Update patch set 2
Patch Set 2:

i've addressed most of the comments.

i've also fixed the code so it compiles. i've tracked the LOG to ALOG change, i've replaced exit(3) calls with dvmAbort calls, i've made -Xint:jit the default for run-test --host, and i've fixed test 063 so it works on the host.

i've added in the missing code cache protection, but it's not as good as the ARM/MIPS version because it leaves the entire tail of the code cache unprotected for the duration of the compilation. it would be nice if a later patch changed things so you compile into a temporary buffer and copy into the cache when you're done, and only unprotect the cache for the duration of the copy.

remaining:

* [code size] cmp $0 -> testl --- can be done in a later patch.
* [performance] movq --- can be done in a later patch.
* [robustness] lea +/- 4 --- can be done in a later patch.

i've tested the x86 JIT on the host, and i've tested that this doesn't seem to have disrupted the ARM/MIPS code.

i think the only blocker here is the NOTICE file, and i'll talk to jbq about that and hopefully get this in today, on the understanding that later patches will fix the problems outlined above.

Patch-set: 2
Label: Code-Review=0
Label: Verified=0