reduce stutter caused by foreground GC on large allocations
foreground GC can block java application for very long time.
It is triggered when the first allocation fails. Sometimes, foreground
GC also fails and heap growth is inevitable. In this situation, we can
skip foreground GC when the requested allocation size is big enough.
The reason to skip the foreground GC is that large allocation will
probably fail until the heap grow.

Change-Id: I0724fda375ea3e3f2684b747d06446ff47b805f5