Update patch set 1
Patch Set 1:

FWIW...

The scheduling was done with a Cortex-A8 in mind.  The A8 is dual-issue, so if two consecutive instructions can operate without hazards they can both issue on the same cycle.

GET_INST_OPCODE and GOTO_OPCODE are both data-processing instructions.  The output is available in the E2 stage, and the inputs are required in E2 or (for the "add" register that is shifted) E1.  So having the AND and ADD directly adjacent will prevent them from issuing on the same clock cycle.

OTOH, LDR doesn't make its result available until E3, and there's no guarantee on how long it will take to complete.  So it's much more important to keep LDR far from any instruction that uses the result.

So +2 on the change, but I wanted to point out that it's not strictly true that there's no reason to split GET_INST_OPCODE and GOTO_OPCODE.

Patch-set: 1
Reviewer: Andy McFadden <1000483@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0