Update patch set 1
Patch Set 1:

We got a nasty bug report on CyanogenMod because of this patch. It appears to cause some insanity in the location framework. Since the patch improves performance so greatly, it would be nice to figure out what the heck is going on.

https://code.google.com/p/cyanogenmod/issues/detail?id=6883

Patch-set: 1
Reviewer: Steve Kondik <1002408@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0