Update patch set 1
Patch Set 1:

Unfortunately mostly doomed.
What you did is good in the sense that it solves the whitespace-in-path issue *inside* the .bat file, where it isn't really an issue to begin with.
However IIRC the main issue is the java command call at the end -- spaces will split paths passed as arguments. Double-quotes are passed literally and won't workaround this problem.

Patch-set: 1