Update patch set 1
Patch Set 1:

I'm not sure how relative paths not working isn't really a problem to begin with. If a tool takes filenames as arguments it should handle relative paths.

On the other hand, people who expect to be able to develop software in paths containing spaces are, IMHO, trying their luck (see GNUMake as a much used example of a tool that can't handle paths with spaces properly).

I will investigate this more when I get time.

Patch-set: 1
Label: Code-Review=0