Update patch set 1
Patch Set 1:

Elliott:
my point here is that toold should not segfault. If user specifies -d but file does not exist by some reason (say, just typo), the original dmtoolsdump segfaults. 

So, I insist that this makes sense :) Probably, if -d argument is not accessible, we could print the error and exit (not warning and continue as I propose), but proposed behaviour is like regular "diff".

Thanks.

Patch-set: 1