Zygote: require seccomp
Refuse to startup zygote if we're running a 3.5 kernel, and
seccomp is not enabled. We'll be doing future work on seccomp,
and third party applications, such as Chrome on Android, need
seccomp for sandboxing.

Seccomp can be enabled in the kernel by setting CONFIG_SECCOMP_FILTER=y

For kernels prior to 3.5, we don't require seccomp, as it would
require backporting work for kernel developers.

Change-Id: I336c2f5fef772bb6dcd26ce517c2b75830fd9da8