Update patch set 2
Patch Set 2:

In that case, I think we have to either reject the whole dex file, or add a new flag for the odex file that indicates the class shouldn't be used.

If we treated it similarly to a normal class verification failure, we would end up writing the class to the odex file as-is, but with no CLASS_ISPREVERIFIED or CLASS_ISOPTIMIZED flags. And then dalvik would attempt to verify the class at runtime, having no knowledge that there was a problem with it during dexopt.

Based on the reasoning that this is indicative of either corruption or tampering, I think it makes sense to reject the dex file as a whole.

Patch-set: 2