Update patch set 2
Patch Set 2:

Hep Happy: Yes, that is correct. For the latest patch set of #57170, lowestSetBit is no longer invoked in Thumb/Gen, since it is invoked in CodegenDriver. You made a comment on that issue. Thanks. A new patch set was created for #57170, and is now merged.
However, your comment is not valid for this patch (#57143).

Patch-set: 2
Reviewer: Anders O Nilsson <1021258@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0