Update patch set 1
Patch Set 1:

i wasn't expecting that we'd actually check this in. i was just uploading it for the next person who has to debug this kind of problem.

we could check it in, i suppose. it would be less odd if we could add it as a hook to just NativeCrypto native methods.

Patch-set: 1
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0