Update patch set 1
Patch Set 1:

It's hard to see the context from only two instructions, but it looks like the two are aliases. In this initial version we didn't handle register aliases, to keep down the complexity. The plan was to improve the alias case further once this initial patch was verified by Google.

The next step was, among other things, to handle some cases of aliasing by replacing the second load instruction with a move, to copy the context from the first register. In other cases we would just use the first register as source when the second one was requested, again eliminating the load instruction.

Patch-set: 1
Reviewer: Henrik Smiding <1022485@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0