Update patch set 1
Patch Set 1: -Code-Review

I'd like to move this Zygote code out of dalvik if possible to the frameworks code, so I'd rather not be adding more dependencies on internals. Yes, I know there will always be some interface between the two, but I'd like it to be more minimal than it is today. When this is pulled out, you'll need to have the code somewhere else, perhaps cutils.

So looking at AttachCurrentThread and dvmAttachCurrentThread, it looks like it does not setThreadName which I guess makes sense, since mucking with the pthread concept of name seems bad for native threads started with pthread_create directly.

Patch-set: 1