Update patch set 1
Patch Set 1:

So the reason for this, is I want to set the main thread of the child to the package name for debugging purposes in the audit records of the device. Right now, it is currently app_process, I want com.bar.foo instead, right justified to the comm task width (16 right now).

Rather then re-implement the code, I came across this, which already does what I want, has been tested and in use, and will reduce my code footprint...win

Patch-set: 1