Update patch set 1
Patch Set 1:

I was holding off on this until we could point out that this needs to be ported to the new art project as well. ART is an alternative to Dalvik now available in KitKat. the art project is in AOSP KitKat branches including kitkat-dev, but still hasn't been pushed to AOSP master. you'll probably want to take a look at it in the meantime. It looks like art's equivalent SetThreadName is already exported and the zygote code is similar enough it should be straightforward to move over. We'll still accept the Dalvik changes in parallel at this time.

Patch-set: 1
Label: Code-Review=0