Update patch set 2
Patch Set 2:

Sorry to be unclear, I was trying to say we'll take the dalvik change too, but we want to take the ART change at the same time so we don't forget. 

I was suggested looking at kitkat-dev for art source because there has been a glitch in getting the kitkat sources into AOSP master which is currently being worked on internally. you could start working on an ART version of this there, but we'll have to cherry-pick it back over to AOSP master when it is available since we can't take commits on kitkat-dev really as I understand it.

Patch-set: 2
Label: Verified=0