Update patch set 1
Patch Set 1:

It's not a JNI function (like NewByteArray), it's a native method called from Java-land.  There's no way to get at the return value -- the move-result instruction won't be executed, and even if it were native-native the convertReferenceResult() function explicitly ignores the return value when an exception is raised.

Patch-set: 1