Update patch set 1
Patch Set 1:

No, I think you are right. I'm making the same mental mistake as the original coder on the other side, they replaced a NewByteArray with a call to this Java method, but didn't consider that they can't use the return value value and need to check for exceptions.

Patch-set: 1