Update patch set 1
Patch Set 1:

This has come up again when rebasing all the clang related changes to current aosp master.

This time I've read the generator more closely to see if there's a better way to do this than to use -Wno-error=unused-function.

My conclusion is that in order to avoid the need for -Wno-error=unused-function, gen-mterp.py would have to be a lot smarter than it currently is.

Currently, it copies all code from header.cpp and then adds CPU specific routines.

The code in header.cpp defines all sorts of functions that may or may not be used by the CPU specific routines, depending on whether or not the CPU specific bits have a better implementation available.

To fix this without -Wno-error=unused-functions, gen-mterp.py would have to actually read the CPU specific bits and based on that throw out parts of header.cpp that aren't needed.

This seems to be too much effort to just avoid having a couple of unused functions in the code that can be thrown out at compile time anyway, especially since this is in dalvik which will likely not see another major revision.

Patch-set: 1
Label: Verified=0