Update patch set 2
Patch Set 2:

vmarko: in cases where you just want a comment fix, it's usually easiest and quickest to make the edit yourself and upload a new patchset. you'll have had to cherrypick the change to test it anyway, so you're halfway there already. saves a lot of round trips, especially when the other person is in a different time zone.

Patch-set: 2