Update patch set 1
Patch Set 1:

First, my apologies for the long delay in giving this CL some attention.  I tested it alone, cherry-picked to yesterday's aosp master and run on a Nexus 10. I tested alone, with aosp master from yesterday, using a command-line version of SciMark2.  FWIW, I saw a nice improvement in Linpack, the LU subtest of SciMark and a possible improvement in DeltaBlue.

Patch-set: 1