Update patch set 1
Patch Set 1:

That's too bad. I would really like to fix this, but unfortunately I'm not in a position to make any modifications due to a change of employment.
Hoisting memory ops outside loops is easily in the top five optimizations for a compiler or JIT. I haven't checked yet, but make sure you do this optimization in ART/LLVM at least.

Patch-set: 1