Update patch set 1
Patch Set 1:

This patch is meant to be used for early 64-bit development and testing and is not meant to be merged as is. It will only work if combined with the patches below:

platform/libcore: https://android-review.googlesource.com/#/c/80449/
platform/build: https://android-review.googlesource.com/#/c/80447/

While these combined patches are expected to provide an experimental version of the Dalvik VM working for both 32-bit and 64-bit, they will break Art, which needs to be modified as well to work with the patched version of libcore.

Also notice that a couple of patches in platform/build may have to be reverted in order for this to work:

https://android-review.googlesource.com/#/c/80526/
https://android-review.googlesource.com/#/c/80233/

Patch-set: 1
Reviewer: Matteo Franchin <1037505@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0