AArch64: Port Dalvik to 64bit
Change Dalvik such that it compiles and runs as a 64bit program.
Supports only the C interpreter and builds only on 64bit hosts.
The VM runs ordinary Dalvik bytecode. The interpreter has been
changed to have 64bit registers, but the Java types remain
unchanged. The garbage collector, JNI and interpreter are
functional.

Change-Id: I8112930a95cb2f8285bb9d285ec102a24651cd1f
Signed-off-by: Stuart Monteith <Stuart.Monteith@arm.com>
Signed-off-by: Dave Butcher <dave.butcher@arm.com>
Signed-off-by: Matteo Franchin <matteo.franchin@arm.com>
Signed-off-by: Vassilis Laganakos<vasileios.laganakos@arm.com>
Signed-off-by: Kévin Petit <kevin.petit@arm.com>