Update patch set 1
Patch Set 1: Code-Review+1

We should be removing the fact that WITH_HOST_DALVIK is disabled for 64-bit builds and be accepting this so we get get the host 64-bit linux builds working.

We removed WITH_HOST_DALVIK in dalvik-dev when we thought we were going to require host dex2oat, but with the compiler performance improvements and problems maintaining a usable darwin build, it stopped doing that.

Even if we have tested and darwin work build is working fine, people doing laptop builds (which tend to be darwin) have wanted this behavior for years, it should be a separate discussion from the 64 bit build whether are going to remove it or not. (or rename to WITH_HOST_ART)

Patch-set: 1