Update patch set 1
Patch Set 1:

If you have a dependency libdvm.so and WITH_HOST_DALVIK is false then something is broken elsewhere. In:
https://android-review.googlesource.com/#/c/97112/
I remove WITH_HOST_DALVIK from ART so that libart.so is available on all host builds. If tests require a runtime then we should switch to libart.

Patch-set: 1