Fixes command line too long error on Windows
On Windows machines the command line can only accept
commands that are 8192 characters or less.  This can
cause a problem when a project has many dependencies
and many libraries are being passed to dex at once.

This fixes it by adding a new option to dex called
--input-list that can read inputs from a file.
This way we only have to pass a single file name
instead of a ton of jar files. This keeps the command
short so Windows will not complain about it.

Change-Id: Ia62ed125c66b24ac81799d3bbac460ae07495b25
Signed-off-by: John Fazekas <jfazekas@fd-software.com>