Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

I don't think we can submit this CL as is, we should split in several parts:
- DexMerge enhancement and usages
- "static" removal and everything that goes in the direction of multi-threading dx (there's also an internal WIP on this)
- Test fixes
- More?

I don't think we need to integrate the BUCK file and the folder permission modifications.
Also, I don't see the point of closing the FileOutputStream instead of the JarOutputStream in the finally block.
And about the Guava files, maybe we could make a real depedency on the guava from the tree?

Patch-set: 2
Reviewer: Benoit Lamarche <1015613@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1