Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Typically debugging log or print messages are left in when there is a feeling they will be valuable later. in art/ the style is to use "if (false)" or some other local "kDebug = false; if (kDebug) ..." to keep them from bit rotting. 

the minority of this is truly dead code removal, that would be fine.

Patch-set: 1
Reviewer: Brian Carlstrom <1003723@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1