Fix for data type mismatch which confuses safemul.
The 64-bit version of dexdump failed with a
"Byte swap + verify failed" error, caused by
the use of data types u4 vs. size_t in a safemul.
On 32-bit this works, but the larger size of size_t
on 64-bit causes a mismatch error.

Bug has now been fixed. Both 32-bit and 64-bit binary
generate the same dexdump output.

This CL also reverses the workaround for the bug
(https://android-review.googlesource.com/107589)
that simply restricted the build to 32-bit only.

Bug: 17442393

Change-Id: Ib51b5d1ecba173f2c2fa2cf884bffa893ce60a53