Update patch set 1
Patch Set 1:

> Using golden files seems fine, but couldn't we use smaller inputs ?
 > Maybe stuff from vm-tests-tf ?

I simply picked my on dex files to start somewhere (the dex size is okay, it is the golden file size that is indeed a bit larger).

Alternatively we could generate the .txt files using a "golden" prebuilt dexdump. Is that easily available to the AOSP tree? The only danger is that bugs can "cancel" eachother out when timed badly....

I will look into the files you suggest also.

Patch-set: 1