Adds an extra compaction step if the wasted space is twice the threshold
For some files, specially now that when merging multiple files only
one pass of compaction is done, an extra compaction might be useful.

In a sample test of a multi way merge, the original merge took 6 seconds
and produced a size of 6106124. The new k-way merge produces, in 1 second,
a file of size 6109516. This changes adds an additional compaction step
and makes the file size even smaller than the original one: 6104752.

Change-Id: I78636d29b25a312ceb7d26ee48c12a040187919e