Update patch set 2
Patch Set 2:

Sounds good to me. I did it because the optimization of the previous CLs yields slightly bigger dex files, and this was because on my 29-way merge it was doing 29 compaction steps. This just adds a second one to remove all the extra space added by variable length encoded stuff. Which in fact gives a better result than the 29 compaction steps, because the last step (the most important one) was not compacted again therefore wasting the variable encoded overhead. If we are happy with this slight increase, we can abandon this CL. We are talking about 5k in a 6MB file.

Patch-set: 2
Reviewer: Esteban de la Canal <1047453@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0