Update patch set 1
Patch Set 1:

> TARGET_OUT_OPTIONAL_EXECUTABLES seems useless?

somewhat. i kind of sort of slightly like the idea that stuff that won't be in /system/bin on a user build isn't there on any build because it's in /system/xbin, but the fact that that's a voluntary manual process weakens it. it's not like there's a CTS test or anything, as this proves.

(i'm assuming here that it's deliberate that dexdump is on user builds, and that the real fix isn't to use the "debug" tag instead.)

Patch-set: 1