Update patch set 2
Patch Set 2: Code-Review-1

I understand the problem here, and that this would be helpful, but I don't think it's a good idea to add a special case to support invalid dex files as input of dexmerger, and to output invalid dex files.
If the ART tests need to test invalid dex files, I don't think it would be reasonable to keep modifying dexmerger to support them.
Wouldn't it be possible not to merge them and use them as separate dex files in the ART tests?
Still, the infinite loop is bad, so ideally I would keep the "if" and throw an AssertionError or a DexException.

Patch-set: 2
Label: Code-Review=-1