Ignore LV debug info when the type doesn't match
As a consequence of the CL that relaxes the conditions to retrieve LV
debug info (455443f47d9e217b99edcc33202f889e2c637b7b), there may be a
SimException because of a "local variable type mismatch".
Some local variable slots may be used for different local variables
with different types, but only one has debug info.
So we need to check if the types match, and ignore the local variable
info if it doesn't.

Bug: 32432143
Test: manual
Change-Id: I0a12ac6476f612c0d8e597c9690587826ac41841