Update patch set 1
Patch Set 1:

> I can't verify today by lack of access to any Windows, I'll try
 > later.
 > 
 > I'm curious, in which context are you using the script?

One of the projects I work on is an Android application that uses Maven. Because of compatibility reasons, I have no option to switch to Gradle. (On the newer projects, Gradle is used by default.)

I know that Maven is not supported for a long time now, but the bug is in this
BAT script, which is part of the SDK's build tools, so I believe it should be fixed.

I have created a minimal example app for testing the issue, it can be found here: https://github.com/GyulaJuhasz/MultiDexScriptExample


Thanks for reviewing the change.

Patch-set: 1