Update patch set 2
Patch Set 2:

I'm not convinced about that.

I agree that it can't really harm but I'd say it's a bad hint for the caller. This synchronize is just preventing corruption of the HashMap but if the synchronize ever protect from a concurrent access it will lead to bad behavior of dx.

It's up to the caller of this clear method to ensure that no dx compilation is running when calling this clear method.

Patch-set: 2
Reviewer: Yohann Roussel <1023201@85c56323-6fa9-3386-8a01-6480fb634889>