Fix local variables with null constant
When a local variable is typed "KNOWN_NULL" because of a null constant,
its type is then transformed to "java.lang.Object", but
"java.lang.Object" wouldn't be added to the types of the dex.
So in the rare case where java.lang.Object wouldn't be referenced
anywhere else in the compilation unit it would fail with
"java.lang.IllegalArgumentException: not found: Ljava/lang/Object;"
Now the associated CstType is "java.lang.Object" so that it is coherent
with the future transformation, and "java.lang.Object" is added to the
list of types.

Test: 136-null-constant-debug-info
Bug: 36824690
Change-Id: I0e7e3057e03761cdb5adbb8f454e3123b3a64969