Update patch set 1
Patch Set 1:

(15 comments)

Liam, your timing is impeccable. Thank you for taking this on.

Support for invoke-custom landed today.

Most of the comments here are nits, but it'd be good to have a test that runs as part of dalvik/tests/run-all-tests. Undoubtedly a bit painful to craft.

Patch-set: 1
Reviewer: Orion Hodson <1071150@85c56323-6fa9-3386-8a01-6480fb634889>