dexdump: use a string instead of str* functions
Clang's static analyzer was complaining that the size in
strncat(buf + bufUsed - 1, param, bufSize - (bufUsed - 1));
was greater than the number of bytes available in the buffer. I don't
see why it thinks that.

At any rate, it's probably better to just use std::string instead of
rolling our own string concatenation/resizing logic. As an added bonus,
this means we'll no longer resize the string on every iteration of the
loop.

There are other places in the file where we use the result of malloc()
without checking it, so I'm assuming crashing on OOM is acceptable here.

Bug: 27101951
Test: mma. Warning is gone.
Change-Id: Id1619ec6d80e20f0c157dbb43d72ac861b8f8064