WIP Tweak dx tests to work when run against libcore
dx tests should be run against OpenJDK core libraries
because that's what dx is compiled against (dx is a
host tool).

Due to a problem, it is currently being run against
libcore; as a workaround, this test changes dex to
avoid using OpenJDK9 covariant overrides that aren't in
libcore. This fixes some failing tests under OpenJDK 9.

Bug: 70525148
Test: dalvik/dx/tests/run-test 041-dex-abstract-method
Change-Id: I8d363a76c6301eef341ce35ab09e56e4ca2454d7