dx: Test script fixes for OpenJDK 9
Add build scripts to each directory and class files generated with
Java 8 compiler. The tests use the latter as golden inputs and remove
the compiler and jasmin as a test running dependency. The build
scripts are only present for updating tests.

There are a few tests where the class files are not added because
there are too many generated during execution. The tests still using
the compiler are 129, 130, 131 and 143.

Bug: 70525148
Test: export EXPERIMENTAL_USE_OPENJDK9=1.8; . build/envsetup.sh; lunch; art/tools/buildbot-build --host -j32; dalvik/dx/tests/run-all-tests --seq
Test: export EXPERIMENTAL_USE_OPENJDK9=false; . build/envsetup.sh; lunch; art/tools/buildbot-build --host -j32; dalvik/dx/tests/run-all-tests --seq

Change-Id: Ifd49e3e81bccb3a0317e9f5677f73d4c5445965e