dx: Fixes for OpenJDK 9
Add build scripts to each directory and class files generated with
Java 8 compiler. The tests use the latter as golden inputs and remove
the compiler as a test running dependency.

There are a projects exceptions where the class files are not added
because there too many (or it's too disruptive): 129, 130, 131 and 143.

Bug: 70525148
Test: EXPERIMENTAL_USE_OPENJDK9=true dx/tests/run-all-tests
Test: EXPERIMENTAL_USE_OPENJDK9=false dx/tests/run-all-tests
Change-Id: Ifd49e3e81bccb3a0317e9f5677f73d4c5445965e