Update patch set 1
Patch Set 1:

PTAL

This change is keep the DX tests alive during the Java 9 migration. It simplifies life not to have the Java compiler as part of testing, but rather to use class files.

There is a resource limit issue running the tests in parallel (the default) that still needs to appear here :/

Patch-set: 1