Build shrinkedAndroid.jar from SDK 20 API
SDK 20 API contains all clases worth considering for multidex tracing
while the latest API contains supplementary classes and is also missing
some.

Also explicitely keep roots in shrinkedAndroid.jar rules. Technically
this is not really needed because they are all kept as part of the
hierarchy of their subclasses (that were already kept), but better be
explicit about the intention.

Test: m shrinkedAndroid
Change-Id: Ifb5ecb8c2803384f52509010bc5f9c3cb5e1791a