Enable -Wimplicit-fallthrough
This adds fallthrough annotations to this code, to make it compile
cleanly when given -Wimplicit-fallthrough. The intent is to enable this
warning for the Android platform after these uses are cleaned up.

I dunno why we don't include android-base here, so I duplicated the
FALLTHROUGH_INTENDED macro. If we can easily depend on android-base,
it's likely slightly better to not duplicate the macro it already
provides.

Bug: 112564944
Test: Builds
Change-Id: I9f49f15d5dd178a2d78a8092d5863a42369968c0